% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{Simulate batch-effect-contaminated miRNA expression data and survival outcome based on the paired datasets from MSKCC.}
\usage{
data_simulation(
  be.surv.data,
  he.data,
  nonzero_beta_position,
  b0value,
  batch_level = "slides",
  t_sort_train = 0,
  t_sort_test = 0,
  t_rev,
  he_train,
  he_test,
  he_train_scale = 1,
  he_test_scale = 1,
  seed = 123
)
}
\arguments{
\item{be.surv.data}{dataset containing true miRNA expression data (i.e. biological effects) and the survival outcome. Expression data starts at the fourth column.}

\item{he.data}{dataset containing pure handling effects and batch variables (e.g. technician, batch, slides).}

\item{nonzero_beta_position}{positions in true \eqn{\beta} that have nonzero values.}

\item{b0value}{vector of nonzero values corresponding to \code{nonzero_beta_position}.}

\item{batch_level}{at what level is batch defined. Either "slides" or "batches". The default is "slides".}

\item{t_sort_train}{0 (default) if training samples are randomly assigned to microarray slides; 1 if training samples are sorted by survival time and then assigned to slides.}

\item{t_sort_test}{0 (default) if test samples are randomly assigned to microarray slides; 1 if test samples are sorted by survival time and then assigned to slides.}

\item{t_rev}{0 if training and test samples are sorted by survival time by the same direction; 1 if sorted by the opposite direction. Ignored if \code{t_sort_train=0} or \code{t_sort_test=0}.}

\item{he_train}{0 if no handling effect in training data; 1 otherwise.}

\item{he_test}{0 if no handling effect in test data; 1 otherwise.}

\item{he_train_scale}{multiplicative scalar applied to the original handling effect for training data (default=1).}

\item{he_test_scale}{multiplicative scalar applied to the original handling effect for test data (default=1).}

\item{seed}{seed for random number generation (default=123).}
}
\value{
A list that contains two datasets and two vectors of batch ids. The first element is the training dataset. The second element is the batch id vector
for the training dataset. The third element is the test dataset. The fourth element is the batch id vector for the test dataset.
}
\description{
\code{data_simulation} performs nonparametric simulation to generate training and test datasets from a pair of well-prepared MSKCC datasets with one containing
true miRNA expression and the other containing pure batch effects. Batch-effect-contaminated miRNA expression data is simulated by virtual hybridization. Survival
outcome is simulated by permutation method. The sizes of the simulated training and test datasets are both 96.
}
\examples{
## Simulate a training dataset and a test dataset with six nonzero true beta values, 
## where both training and test data contain batch effects 
## that are not associated with the survival outcome.
sim1=data_simulation(be.survival, he, nonzero_beta_position=c(385,866,1010,2218,2660,3026), 
b0value=c(1.04,1.40,1.45,1.62,3.13,1.76), t_sort_train=0, t_sort_test=0, 
t_rev=0, he_train=1, he_test=1, he_train_scale=1, he_test_scale=1, seed=123)
}
\references{
Ni, A., Qin, L. (2021) "Performance Evaluation of Transcriptomics Data Normalization for Survival Risk Prediction". \emph{Briefings in Bioinformatics}, 22(6), bbab257.
}
