% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.simulation.R
\name{survival.simulation}
\alias{survival.simulation}
\title{Comparison of prediction accuracy of normalization methods in survival analysis.}
\usage{
survival.simulation(
  be.surv.data,
  he.data,
  nonzero_position,
  b0value,
  ps_rho_cutoff = 0.9,
  t_sort_train = 0,
  t_sort_test = 0,
  t_rev,
  he_train,
  he_test,
  he_train_scale = 1,
  he_test_scale = 1,
  norm_type,
  norm_train,
  norm_test,
  univ_cutoff_range = c(0, 0.05),
  nuniv_cutoff = 20,
  lambda_max_glmnet = 0.3,
  nlambda = 30,
  nfold = 6,
  nsim = 400,
  seed = 123
)
}
\arguments{
\item{be.surv.data}{dataset containing true miRNA expression data (i.e. biological effects) and the survival outcome.}

\item{he.data}{dataset containing pure handling effects and batch variables (e.g. technician, batch, slides).}

\item{nonzero_position}{positions in true \eqn{\beta} that have nonzero values.}

\item{b0value}{vector of nonzero values corresponding to \code{nonzero_position}.}

\item{ps_rho_cutoff}{cutoff of correlation coefficient during pre-screen (default=0.9).}

\item{t_sort_train}{0 (default) if training samples are randomly assigned to microarray slides; 1 if training samples are sorted by survival time and then assigned to slides.}

\item{t_sort_test}{0 (default) if test samples are randomly assigned to microarray slides; 1 if test samples are sorted by survival time and then assigned to slides.}

\item{t_rev}{0 if training and test samples are sorted by survival time by the same direction; 1 if sorted by the opposite direction. Ignored if \code{t_sort_train=0} or \code{t_sort_test=0}.}

\item{he_train}{0 if no handling effect in training data; 1 otherwise.}

\item{he_test}{0 if no handling effect in test data; 1 otherwise.}

\item{he_train_scale}{multiplicative scalar applied to the original handling effect for training data (default=1).}

\item{he_test_scale}{multiplicative scalar applied to the original handling effect for test data (default=1).}

\item{norm_type}{normalization methods (0: no normalization; 1: median normalization; 2: quantile normalization; 3: Variance stabilizing transformation (VSN).}

\item{norm_train}{0 if no normalization applied to training data; 1 otherwise. Ignored if \code{norm_type=0}.}

\item{norm_test}{0 if no normalization applied to test data; 1 otherwise. Ignored if \code{norm_type=0}.}

\item{univ_cutoff_range}{range of cutoffs of p values for univariate analysis (default=\code{c(0,0.05)}).}

\item{nuniv_cutoff}{number of equal-spaced cutoff values in \code{univ_cutoff_range} (default=20).}

\item{lambda_max_glmnet}{maximum value of grid-search range of tuning parameter lambda in lasso and adaptive lasso (default=0.3).}

\item{nlambda}{number of equal-sapced lambda values in \code{lambda_max_glmnet} (default=30).}

\item{nfold}{number of folds for the cross-validation for tuning parameter selection in lasso and adaptive lass (default=6).}

\item{nsim}{number of simulation replicates (default=400).}

\item{seed}{seed for random number generation (default=123).}
}
\value{
An object of class \code{sim.result} containing estimated regression coefficients and C-indices from each method.
\item{bhat_oracle}{estimated regression coefficients from Oracle method (each row represents one simulation replicate; each column represents one marker)}
\item{bhat_univ}{estimated regression coefficients from univariate method}
\item{bhat_lasso}{estimated regression coefficients from lasso method}
\item{bhat_alasso}{estimated regression coefficients from adaptive lasso method}
\item{c_stats}{C-indices from all four methods based on test data (each column represents one method)}
\item{lambdas}{selected tuning parameter lambda for Lasso and adaptive Lasso methods}
\item{subset_size}{the sizes of data used for variable selection after pre-screening}
}
\description{
\code{simulation} performs nonparametric simulation to evaluate prediction accuracy of three normalization methods (median normalization, quantile normalization,
Variance stabilizing transformation) using a pair of well-prepared datasets with one containing true miRNA expression and the other containing pure handling effects.
}
\examples{
## conduct a simulation with six nonzero true beta values, 
## where both training and test data contain handling effects 
## that are not associated with the survival outcome, and both 
## are subject to median normalization.
sim1=simulation(be.surv, he, nonzero_position=c(385,866,1010,2218,2660,3026), 
b0value=c(1.04,1.40,1.45,1.62,3.13,1.76), t_sort_train=0, t_sort_test=0, 
t_rev=0, he_train=1, he_test=1, norm_type=1, norm_train=1, norm_test=1)
}
\references{
Ni, A., Qin, L. (2021+) "Performance Evaluation of Transcriptomics Data Normalization for Survival Risk Prediction". accepted by \emph{Briefings in Bioinformatics}.
}
